added .lean to services

+++ controller/users.js
@@ -18,18 +18,30 @@ const getEmails = async (req, res) => {
     res.json(await user.getEmails());
 }
 const updateUser = async (req, res) => {
+    if (req.params.userId !== req.params.id) {
+        return res.status(500).json({errors: ["unable to update, requester is not the user"]});
+    }
     res.json(await user.updateUser(req.params
         .id, req.params.email, req.params.firstName, req.params.lastName, req.params.password));
 }
 const updateUserImg = async (req, res) => {
+    if (req.params.userId !== req.params.id) {
+        return res.status(500).json({errors: ["unable to update, requester is not the user"]});
+    }
     res.json(await user.updateUserImg(req.params.id, req.params.img));
 }
 const deleteUser = async (req, res) => {
+    if (req.params.userId !== req.params.id) {
+        return res.status(500).json({errors: ["unable to delete, requester is not the user"]});
+    }
     await like.removeLikesByUser(req.params.id);
     await comment.deleteCommentsByUser(req.params.id)
     res.json(await user.deleteUser(req.params.id));
 }
 const deleteUserByEmail = async (req, res) => {
+    if (req.params.userId !== req.params.id) {
+        return res.status(500).json({errors: ["unable to delete, requester is not the user"]});
+    }
     await like.removeLikesByUser(req.params.id);
     await comment.deleteCommentsByUser(req.params.id)
     res.json(await user.deleteUserByEmail(req.params.email));
